Adds option to treat unknown cells as empty in signed_distance_field.
This CL adds an option when creating a signed distance field object from
an occupancy grid measurement to treat unknown cells as empty cells.
This option is defaulted to true. Setting this option to false causes
unknown cells to be treated as obstacles. This CL also updates the test
occupancy grid with a region of unknown cells.