Uses trajectory status for reinit decision in primitive_planner.
This commit updates the primitive_planner to use the status information
from the last published trajectory as part of its reinitialization
criteria. This allows the primitive planner to diferrentiate between
cases where the robot is stopped and no continuity is expected and cases
where the robot is temporarily stationary in the middle of a planned
maneuver.